Setup OCR link to form fill
I setup the link but there are still a few issues. While it fills some fields, certain fields such as the date field aren't filled even when a date is provided. This most likely has to do with the AI providing a JSON file for the data but the form isn't able to take that JSON data.